refactor(tools): replace sms_wait_for_code with sms_list_messages
A blocking 90s wait tool hid failures and could hang forever. Mirror
email_list_messages: one GET, agent sleeps and lists, optional claim.
