+++ src/core/agents/offSecAgent/tools/index.ts
@@ -124,11 +124,11 @@ export {
   ScopeViolationError,
 } from "./scopeGuard";
 export {
+  SMS_LIST_MESSAGES_TOOL_NAME,
   SMS_TOOL_NAMES,
-  SMS_WAIT_FOR_CODE_TOOL_NAME,
   sessionHasSmsPasswordless,
-  smsWaitForCode,
-} from "./smsWaitForCode";
+  smsListMessages,
+} from "./smsListMessages";
 export { spawnCodingAgent } from "./spawnCodingAgent";
 export { spawnPentestAgent } from "./spawnPentestAgent";
 export { spawnPentestSwarm } from "./spawnPentestSwarm";
@@ -221,7 +221,7 @@ import { runAttackSurface } from "./runAttackSurface";
 import { runCodeQuery } from "./runCodeQuery";
 import { runPentestWorkflow } from "./runPentestWorkflow";
 import { runWhiteboxScan } from "./runWhiteboxScan";
-import { smsWaitForCode } from "./smsWaitForCode";
+import { smsListMessages } from "./smsListMessages";
 import { spawnCodingAgent } from "./spawnCodingAgent";
 import { spawnPentestAgent } from "./spawnPentestAgent";
 import { spawnPentestSwarm } from "./spawnPentestSwarm";
@@ -348,8 +348,8 @@ export function createAllTools(ctx: ToolContext) {
     email_search_messages: emailSearchMessages(ctx),
     email_get_message: emailGetMessage(ctx),
 
-    // Inbound SMS OTP wait (gated at activeTools level when no Mobile OTP cred)
-    sms_wait_for_code: smsWaitForCode(ctx),
+    // Inbound SMS list (gated at activeTools level when no Mobile OTP cred)
+    sms_list_messages: smsListMessages(ctx),
 
     // Web search tools (requires Pensar account)
     web_search: webSearch(ctx),
@@ -469,7 +469,7 @@ export const ALL_TOOL_NAMES: ToolName[] = [
   "email_get_attachments",
   "email_mark_read",
   "send_email",
-  "sms_wait_for_code",
+  "sms_list_messages",
   // Web search (requires Pensar account)
   "web_search",
   "get_page",
@@ -586,5 +586,5 @@ export const SKILL_TOOL_NAMES = ["read_skill"] as const;
 /** Email inbox tool names — filtered out by the base class when no inboxes are configured. */
 export { EMAIL_TOOL_NAMES as EMAIL_TOOL_NAMES_ACTIVE } from "./email";
 
-/** SMS wait tool names — filtered out by the base class when no Mobile OTP credential is present. */
-export { SMS_TOOL_NAMES as SMS_TOOL_NAMES_ACTIVE } from "./smsWaitForCode";
+/** SMS list tool names — filtered out by the base class when no Mobile OTP credential is present. */
+export { SMS_TOOL_NAMES as SMS_TOOL_NAMES_ACTIVE } from "./smsListMessages";
