+++ src/core/agents/offSecAgent/tools/smsListMessages.test.ts
@@ -2,10 +2,10 @@ import { afterEach, describe, expect, it, vi } from "vitest";
 import { CredentialManager } from "../../../credentials";
 import type { SessionInfo } from "../../../session";
 import {
-  SMS_WAIT_FOR_CODE_TOOL_NAME,
+  SMS_LIST_MESSAGES_TOOL_NAME,
   sessionHasSmsPasswordless,
-  smsWaitForCode,
-} from "./smsWaitForCode";
+  smsListMessages,
+} from "./smsListMessages";
 import type { ToolContext } from "./types";
 
 const executeOpts = { toolCallId: "call_1", messages: [] };
@@ -19,12 +19,43 @@ function makeCtx(cm?: CredentialManager): ToolContext {
   };
 }
 
+function listedResponse() {
+  return {
+    ok: true,
+    status: 200,
+    json: async () => ({
+      messages: [
+        {
+          id: "msg-1",
+          body: "Your code is 424242",
+          code: "424242",
+          fromPhoneNumber: "+15550001111",
+          toPhoneNumber: "+15551234567",
+          receivedAt: "2026-04-10T12:00:00.000Z",
+          consumedAt: null,
+        },
+      ],
+      claimed: null,
+    }),
+  };
+}
+
 function claimedResponse() {
   return {
     ok: true,
     status: 200,
     json: async () => ({
-      messages: [],
+      messages: [
+        {
+          id: "msg-1",
+          body: "Your code is 424242",
+          code: "424242",
+          fromPhoneNumber: "+15550001111",
+          toPhoneNumber: "+15551234567",
+          receivedAt: "2026-04-10T12:00:00.000Z",
+          consumedAt: "2026-04-10T12:00:01.000Z",
+        },
+      ],
       claimed: {
         id: "msg-1",
         body: "Your code is 424242",
@@ -67,34 +98,33 @@ describe("sessionHasSmsPasswordless", () => {
   });
 });
 
-describe("smsWaitForCode", () => {
+describe("smsListMessages", () => {
   afterEach(() => {
     vi.unstubAllGlobals();
     vi.unstubAllEnvs();
-    vi.useRealTimers();
   });
 
   it("uses the stable tool name", () => {
-    expect(SMS_WAIT_FOR_CODE_TOOL_NAME).toBe("sms_wait_for_code");
+    expect(SMS_LIST_MESSAGES_TOOL_NAME).toBe("sms_list_messages");
   });
 
   it("fails loud when AGENT_API_URL is unset", async () => {
     vi.stubEnv("AGENT_API_URL", "");
     vi.stubEnv("AGENT_API_TOKEN", "token");
-    const tool = smsWaitForCode(makeCtx());
+    const tool = smsListMessages(makeCtx());
     await expect(
       tool.execute?.(
         {
           sinceMs: 1,
           phoneNumber: "+15551234567",
-          toolCallDescription: "wait",
+          toolCallDescription: "list",
         },
         executeOpts,
       ),
     ).rejects.toThrow("AGENT_API_URL");
   });
 
-  it("claims the first inbound SMS and returns the parsed code", async () => {
+  it("lists inbound SMS without claiming", async () => {
     vi.stubEnv("AGENT_API_URL", "https://api.example.com");
     vi.stubEnv("AGENT_API_TOKEN", "token");
     const cm = new CredentialManager();
@@ -104,35 +134,68 @@ describe("smsWaitForCode", () => {
       loginUrl: "https://app.example.com/login",
     });
 
-    const fetchMock = vi.fn().mockResolvedValue(claimedResponse());
+    const fetchMock = vi.fn().mockResolvedValue(listedResponse());
     vi.stubGlobal("fetch", fetchMock);
 
-    const tool = smsWaitForCode(makeCtx(cm));
+    const tool = smsListMessages(makeCtx(cm));
     const result = await tool.execute?.(
       {
         credentialId: "cred-1",
         sinceMs: 1_700_000_000_000,
-        toolCallDescription: "wait for login code",
+        toolCallDescription: "list inbound SMS",
       },
       executeOpts,
     );
 
     expect(result).toEqual({
       success: true,
-      code: "424242",
-      body: "Your code is 424242",
-      fromPhoneNumber: "+15550001111",
-      receivedAt: "2026-04-10T12:00:00.000Z",
-      messageId: "msg-1",
+      messages: [
+        {
+          id: "msg-1",
+          body: "Your code is 424242",
+          code: "424242",
+          fromPhoneNumber: "+15550001111",
+          toPhoneNumber: "+15551234567",
+          receivedAt: "2026-04-10T12:00:00.000Z",
+          consumedAt: null,
+        },
+      ],
+      claimed: null,
     });
     expect(fetchMock).toHaveBeenCalledTimes(1);
     const url = String(fetchMock.mock.calls[0]?.[0]);
     expect(url).toContain("/agent/sms/messages?");
     expect(url).toContain("toPhoneNumber=%2B15551234567");
+    expect(url).not.toContain("claim=");
+  });
+
+  it("claims the newest unconsumed SMS when claim is true", async () => {
+    vi.stubEnv("AGENT_API_URL", "https://api.example.com");
+    vi.stubEnv("AGENT_API_TOKEN", "token");
+
+    const fetchMock = vi.fn().mockResolvedValue(claimedResponse());
+    vi.stubGlobal("fetch", fetchMock);
+
+    const tool = smsListMessages(makeCtx());
+    const result = await tool.execute?.(
+      {
+        phoneNumber: "+15551234567",
+        sinceMs: 1,
+        claim: true,
+        toolCallDescription: "claim login code",
+      },
+      executeOpts,
+    );
+
+    expect(result).toMatchObject({
+      success: true,
+      claimed: { id: "msg-1", code: "424242" },
+    });
+    const url = String(fetchMock.mock.calls[0]?.[0]);
     expect(url).toContain("claim=1");
   });
 
-  it("returns a lease conflict without retrying forever", async () => {
+  it("returns a lease conflict without retrying", async () => {
     vi.stubEnv("AGENT_API_URL", "https://api.example.com");
     vi.stubEnv("AGENT_API_TOKEN", "token");
 
@@ -143,12 +206,12 @@ describe("smsWaitForCode", () => {
     });
     vi.stubGlobal("fetch", fetchMock);
 
-    const tool = smsWaitForCode(makeCtx());
+    const tool = smsListMessages(makeCtx());
     const result = await tool.execute?.(
       {
         phoneNumber: "+15551234567",
         sinceMs: 1,
-        toolCallDescription: "wait",
+        toolCallDescription: "list",
       },
       executeOpts,
     );
@@ -161,80 +224,33 @@ describe("smsWaitForCode", () => {
     expect(fetchMock).toHaveBeenCalledTimes(1);
   });
 
-  it("keeps polling after a 409 already-consumed race, then claims the next message", async () => {
+  it("returns already-consumed without polling", async () => {
     vi.stubEnv("AGENT_API_URL", "https://api.example.com");
     vi.stubEnv("AGENT_API_TOKEN", "token");
-    vi.useFakeTimers();
 
-    const fetchMock = vi
-      .fn()
-      .mockResolvedValueOnce({
-        ok: false,
-        status: 409,
-        statusText: "Conflict",
-      })
-      .mockResolvedValueOnce(claimedResponse());
-    vi.stubGlobal("fetch", fetchMock);
-
-    const tool = smsWaitForCode(makeCtx());
-    const pending = tool.execute?.(
-      {
-        phoneNumber: "+15551234567",
-        sinceMs: 1,
-        toolCallDescription: "wait",
-      },
-      executeOpts,
-    );
-
-    await vi.advanceTimersByTimeAsync(2_000);
-    const result = await pending;
-
-    expect(result).toMatchObject({ success: true, code: "424242" });
-    expect(fetchMock).toHaveBeenCalledTimes(2);
-  });
-
-  it("releases the lease after timeout", async () => {
-    vi.stubEnv("AGENT_API_URL", "https://api.example.com");
-    vi.stubEnv("AGENT_API_TOKEN", "token");
-    vi.useFakeTimers();
-
-    const fetchMock = vi.fn().mockImplementation(async (url: string) => {
-      if (String(url).includes("/agent/sms/lease/release")) {
-        return {
-          ok: true,
-          status: 200,
-          json: async () => ({ released: true }),
-        };
-      }
-      return {
-        ok: true,
-        status: 200,
-        json: async () => ({ claimed: null, messages: [] }),
-      };
+    const fetchMock = vi.fn().mockResolvedValue({
+      ok: false,
+      status: 409,
+      statusText: "Conflict",
     });
     vi.stubGlobal("fetch", fetchMock);
 
-    const tool = smsWaitForCode(makeCtx());
-    const pending = tool.execute?.(
+    const tool = smsListMessages(makeCtx());
+    const result = await tool.execute?.(
       {
         phoneNumber: "+15551234567",
         sinceMs: 1,
-        toolCallDescription: "wait",
+        claim: true,
+        toolCallDescription: "claim",
       },
       executeOpts,
     );
 
-    await vi.advanceTimersByTimeAsync(95_000);
-    const result = await pending;
-
-    expect(result).toMatchObject({
+    expect(result).toEqual({
       success: false,
-      error: expect.stringContaining("Timed out after 90s"),
+      error:
+        "That SMS was already claimed. List without claim, or wait for a newer message.",
     });
-    expect(
-      fetchMock.mock.calls.some((call) =>
-        String(call[0]).includes("/agent/sms/lease/release"),
-      ),
-    ).toBe(true);
+    expect(fetchMock).toHaveBeenCalledTimes(1);
   });
 });
