+++ src/core/agents/specialized/authenticationAgent/agent.ts
@@ -209,8 +209,8 @@ export class AuthenticationAgent extends OffensiveSecurityAgent<AuthenticationRe
         "email_get_message",
         // Send email (filtered out by base class when no SMTP configured)
         "send_email",
-        // Mobile OTP wait (filtered out by base class when no sms-passwordless cred)
-        "sms_wait_for_code",
+        // Mobile OTP list (filtered out by base class when no sms-passwordless cred)
+        "sms_list_messages",
         // Web search tools — look up auth bypass techniques, default credentials
         "web_search",
         "get_page",
@@ -329,9 +329,11 @@ function buildAuthPrompt(
     const smsInstructions = hasSmsPasswordless
       ? `
 If a credential has a phoneNumber additional field (Mobile OTP / sms-passwordless), phone is the login
-identifier — not MFA-after-password. browser_fill credentialField="phoneNumber", click send-code, call
-sms_wait_for_code with sinceMs = Date.now() at that click, then browser_fill the OTP. Do NOT report
-phone_verification as a barrier for this flow. TOTP-via-env for authenticator MFA is unchanged.`
+identifier — not MFA-after-password. browser_fill credentialField="phoneNumber", click send-code, then
+execute_command \`sleep 5\` and call sms_list_messages with sinceMs = Date.now() at that click (claim=true
+once a message is present). If the list is empty, sleep and list again; if it stays empty, fail with what
+you observed — do not hang waiting. Then browser_fill the OTP. Do NOT report phone_verification as a
+barrier for this flow. TOTP-via-env for authenticator MFA is unchanged.`
       : "";
     parts.push(`INSTRUCTIONS:
 You have credentials available via credential IDs — authenticate immediately.
