+++ src/core/agents/specialized/pentest/agent.test.ts
@@ -493,7 +493,7 @@ The following credentials are available. To use a credential, pass its ID to the
     expect(prompt).toContain("Standard user account for the customer portal");
     // And it's told to report_error rather than silently skip if auth fails.
     expect(prompt).toContain("authentication_failed");
-    expect(prompt).toContain("sms_wait_for_code");
+    expect(prompt).toContain("sms_list_messages");
   });
 
   it("points the agent at the credential Context instructions instead of defaulting to a browser login", () => {
