+++ src/core/agents/specialized/pentest/agent.ts
@@ -1124,7 +1124,7 @@ Do NOT discover or enumerate other endpoints or services. Focus exclusively on t
   // spawned worker receive the operator's auth instructions; when a live session
   // is also present, these double as re-auth guidance if it expires.
   const credentialSection = credentialContext
-    ? `\n## Available Credentials\nThe operator provided the following credentials and authentication instructions for this engagement. Authenticate by following the instructions in each credential's Context exactly — use the method it describes (for example, a token/API exchange via execute_command or http_request) rather than defaulting to a browser login. Treat the Context as the source of truth for how to authenticate, and how to re-authenticate if a provided session expires. When a tool needs a secret value and supports it (e.g. browser_fill), reference it by credentialId + credentialField so the secret resolves securely at execution time instead of hardcoding it. If you cannot authenticate with these, call report_error with reason "authentication_failed" and a specific message rather than reporting a non-finding.\nWhen a credential has additional field phoneNumber or authMethod sms-passwordless, phone is the login identifier (not MFA-after-password). Fill credentialField="phoneNumber", click send-code, call sms_wait_for_code with sinceMs from that click, then fill the OTP. Do not report phone_verification as a barrier. TOTP-via-environment-variable for authenticator MFA is unchanged.\n\n${credentialContext}\n`
+    ? `\n## Available Credentials\nThe operator provided the following credentials and authentication instructions for this engagement. Authenticate by following the instructions in each credential's Context exactly — use the method it describes (for example, a token/API exchange via execute_command or http_request) rather than defaulting to a browser login. Treat the Context as the source of truth for how to authenticate, and how to re-authenticate if a provided session expires. When a tool needs a secret value and supports it (e.g. browser_fill), reference it by credentialId + credentialField so the secret resolves securely at execution time instead of hardcoding it. If you cannot authenticate with these, call report_error with reason "authentication_failed" and a specific message rather than reporting a non-finding.\nWhen a credential has additional field phoneNumber or authMethod sms-passwordless, phone is the login identifier (not MFA-after-password). Fill credentialField="phoneNumber", click send-code, sleep with execute_command, then sms_list_messages with sinceMs from that click (claim=true once a message is present). If the list stays empty, stop and report — do not hang. Do not report phone_verification as a barrier. TOTP-via-environment-variable for authenticator MFA is unchanged.\n\n${credentialContext}\n`
     : "";
 
   const contextSection = context
@@ -1229,8 +1229,8 @@ const SHARED_PENTEST_TOOLS = [
   "email_get_message",
   // Send email (filtered out by base class when no SMTP configured)
   "send_email",
-  // Mobile OTP wait (filtered out by base class when no sms-passwordless cred)
-  "sms_wait_for_code",
+  // Mobile OTP list (filtered out by base class when no sms-passwordless cred)
+  "sms_list_messages",
   // Memory tools — for on-demand retrieval of project knowledge
   "list_memories",
   "get_memory",
